api/requirements.txt
Flask >= 2.0.1
PyJWT
mysql_connector >= 2.2.9
opencv-python
flask-cors
python-dotenv